% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luchtmeetnet.R
\name{luchtmeetnet_get_data}
\alias{luchtmeetnet_get_data}
\title{R interface to luchtmeetnet api}
\usage{
luchtmeetnet_get_data(obj, params = list(), verbose = FALSE)
}
\arguments{
\item{obj}{The object to query}

\item{params}{The params to pass to the query}

\item{verbose}{Whether to print out stuffs}
}
\value{
A data.frame with reponse data
}
\description{
R interface to luchtmeetnet api
}
