% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report.R
\name{luchtmeetnet_get_report}
\alias{luchtmeetnet_get_report}
\title{R wrapper for luchtmeetnet report api}
\usage{
luchtmeetnet_get_report(components, report, start, end, verbose = FALSE)
}
\arguments{
\item{components}{List of components (pm25, pm10, no2, o3)}

\item{report}{The report (hourly)}

\item{start}{Start date (i.e. 2014-01-01T20:30:00Z)}

\item{end}{End date (i.e. 2014-01-01T20:30:00Z)}

\item{verbose}{Whether to print out stuffs}

\item{obj}{The object to query}
}
\value{
A data.frame with reponse data
}
\description{
R wrapper for luchtmeetnet report api
}
